refactor: simplify and split into multiple files
