+++ addrs.go
@@ -1,0 +1,58 @@
+package websocket
+
+import (
+	"fmt"
+	"net"
+	"net/url"
+
+	ma "github.com/multiformats/go-multiaddr"
+	manet "github.com/multiformats/go-multiaddr-net"
+	ws "golang.org/x/net/websocket"
+)
+
+func ConvertWebsocketMultiaddrToNetAddr(maddr ma.Multiaddr) (net.Addr, error) {
+	_, host, err := manet.DialArgs(maddr)
+	if err != nil {
+		return nil, err
+	}
+
+	a := &ws.Addr{
+		URL: &url.URL{
+			Host: host,
+		},
+	}
+	return a, nil
+}
+
+func ParseWebsocketNetAddr(a net.Addr) (ma.Multiaddr, error) {
+	wsa, ok := a.(*ws.Addr)
+	if !ok {
+		return nil, fmt.Errorf("not a websocket address")
+	}
+
+	tcpaddr, err := net.ResolveTCPAddr("tcp", wsa.Host)
+	if err != nil {
+		return nil, err
+	}
+
+	tcpma, err := manet.FromNetAddr(tcpaddr)
+	if err != nil {
+		return nil, err
+	}
+
+	wsma, err := ma.NewMultiaddr("/ws")
+	if err != nil {
+		return nil, err
+	}
+
+	return tcpma.Encapsulate(wsma), nil
+}
+
+func parseMultiaddr(a ma.Multiaddr) (string, error) {
+	_, host, err := manet.DialArgs(a)
+	if err != nil {
+		return "", err
+	}
+
+	return "ws://" + host, nil
+}
