+++ dialer.go
@@ -1,0 +1,41 @@
+package websocket
+
+import (
+	"context"
+
+	ws "github.com/gorilla/websocket"
+	tpt "github.com/libp2p/go-libp2p-transport"
+	ma "github.com/multiformats/go-multiaddr"
+	manet "github.com/multiformats/go-multiaddr-net"
+)
+
+type dialer struct{}
+
+func (d *dialer) Dial(raddr ma.Multiaddr) (tpt.Conn, error) {
+	return d.DialContext(context.Background(), raddr)
+}
+
+func (d *dialer) DialContext(ctx context.Context, raddr ma.Multiaddr) (tpt.Conn, error) {
+	wsurl, err := parseMultiaddr(raddr)
+	if err != nil {
+		return nil, err
+	}
+
+	wscon, _, err := ws.DefaultDialer.Dial(wsurl, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	mnc, err := manet.WrapNetConn(NewConn(wscon, nil))
+	if err != nil {
+		return nil, err
+	}
+
+	return &wsConn{
+		Conn: mnc,
+	}, nil
+}
+
+func (d *dialer) Matches(a ma.Multiaddr) bool {
+	return WsFmt.Matches(a)
+}
