+++ listener.go
@@ -1,0 +1,73 @@
+package websocket
+
+import (
+	"context"
+	"fmt"
+	"net/http"
+	"net/url"
+
+	tpt "github.com/libp2p/go-libp2p-transport"
+	ma "github.com/multiformats/go-multiaddr"
+	manet "github.com/multiformats/go-multiaddr-net"
+)
+
+type wsConn struct {
+	manet.Conn
+	t tpt.Transport
+}
+
+var _ tpt.Conn = (*wsConn)(nil)
+
+func (c *wsConn) Transport() tpt.Transport {
+	return c.t
+}
+
+type listener struct {
+	manet.Listener
+
+	incoming chan *Conn
+
+	tpt tpt.Transport
+
+	origin *url.URL
+}
+
+func (l *listener) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	c, err := upgrader.Upgrade(w, r, nil)
+	if err != nil {
+		http.Error(w, "Failed to upgrade websocket", 400)
+		return
+	}
+
+	ctx, cancel := context.WithCancel(context.Background())
+	l.incoming <- NewConn(c, cancel)
+
+	// wait until conn gets closed, otherwise the handler closes it early
+	<-ctx.Done()
+}
+
+func (l *listener) Accept() (tpt.Conn, error) {
+	c, ok := <-l.incoming
+	if !ok {
+		return nil, fmt.Errorf("listener is closed")
+	}
+
+	mnc, err := manet.WrapNetConn(c)
+	if err != nil {
+		return nil, err
+	}
+
+	return &wsConn{
+		Conn: mnc,
+		t:    l.tpt,
+	}, nil
+}
+
+func (l *listener) Multiaddr() ma.Multiaddr {
+	wsma, err := ma.NewMultiaddr("/ws")
+	if err != nil {
+		panic(err)
+	}
+
+	return l.Listener.Multiaddr().Encapsulate(wsma)
+}
